fix user id authorization

+++ src/service/contactService.js
@@ -1,11 +1,9 @@
 
 const { Contact } = require('../db/contactModel')
-// const select = '-_id -owner -__v'
 const { NotAuthorized } = require('../helpers/errors')
 
 const getContact = async (userId) => {
-  console.log('userId', userId)
-  return await Contact.find({ owner: userId })
+  return await Contact.find({ owner: userId }).select({ __v: 0 })
 }
 const getContactById = async (userId, id) => {
   if (!userId) {
@@ -13,15 +11,13 @@ const getContactById = async (userId, id) => {
   }
   return await Contact.findById({ owner: userId, _id: id })
 }
-const addContact = async ({ name, email, phone }, userId) => {
-  console.log('userId', userId)
+const addContact = async ({ name, email, phone, userId }) => {
   const newClient = new Contact({
     name,
     email,
     phone,
     owner: userId,
   })
-  console.log('newClient', newClient)
   return await newClient.save()
 }
 
@@ -38,8 +34,8 @@ const updateStatusContact = async (id, userId, { favorite }) => {
   )
   return updateClient
 }
-const deleteContact = async (id) => {
-  return await Contact.findByIdAndRemove(id)
+const deleteContact = async (id, userId) => {
+  return await Contact.findByIdAndRemove({ _id: id, owner: userId })
 }
 
 module.exports = {
